New object building  primitives and two new mixing models for fractal box, bug fixes
Introduced object primitives #ellipsoid and #cone. The #ellipsoid works with regular semi-axes
along x, y, and z. The #cone works in a similar way to the cylinder and can be truncated with circular faces of different radii. If one radius is zero then it is a normal cone.
Introduced two new mixing models for using with a fractal box, #material_range and #material_list
The #material_range allows materials to be constructed using ranges of permittivity, conductivity, magnetic permeability and magnetic loss. Ranges can be increasing or decreasing as needed.
The #material_list creates a list of predefined materials that can be used with a fractal box. The order of the listed materials reflects the order of the bins they are allocated to in a fractal box.
The fractal box and associated mixing models have been updated to use a list of material IDs and do not rely on the assumption that are generated sequentially. This gives greater flexibility and
adds little computation burden on the creation of the fractal volume.

Bug fixes for mixing anisotropic volumetric materials where "mr" was used instead of "ms".
Bug fixes on objects that use integer coordinates for centres but where passed floats instead.