Move materials back into FDTDGrid class
Materials are used to by the grid to build PMLs and components as well
as creating new materials to voltage sources. Additionally, didn't parts
of the grid will need a different subset of materials. Storing in Grid
could in the future reduce data duplication and the duplicate
computation looks to be minimal.