Remove logger.exception from outside try except statement
logger.exception should only be used inside a try except statement where
there is some error/stacktrace that needs outputting as well as the
error message. Instead, should use logger.error then raise an error,
and/or pass the message to the error being raised (which the calling
function should handle and/or log the exception).